ritm_service/requirements.txt
opencv-python
Flask==2.2.2
requests-toolbelt==0.9.1
scipy
numpy
Cython
scikit-image
Pillow
matplotlib
imgaug>=0.4
albumentations>0.5
onnxruntime-gpu
graphviz
tqdm
pyyaml
easydict
torch>=1.4.0
torchvision>=0.5.0
tensorboard==2.10.0
future
cffi
ninja
protobuf==3.19.4
aqueduct==1.10.12